+++ src/gui/ScoreDialog.java
@@ -71,7 +71,7 @@ public class ScoreDialog
         m_areaWhite.setText(Integer.toString(score.m_areaWhite));
         m_capturedBlack.setText(Integer.toString(score.m_capturedBlack));
         m_capturedWhite.setText(Integer.toString(score.m_capturedWhite));
-        m_komi.setText(Float.toString(score.m_komi));
+        m_komi.setText(Double.toString(score.m_komi));
         m_resultChinese.setText(Score.formatResult(score.m_resultChinese));
         m_resultJapanese.setText(Score.formatResult(score.m_resultJapanese));
         m_rules.setText(score.m_rules == go.Board.RULES_JAPANESE ? "Japanese"
